requirements_load.txt
requests
python-dotenv
psycopg2-binary
sqlalchemy
psycopg2